Updating minRangeRequestSize to take long instead of int (#804)
Updating minRangeRequestSize to take long instead of int

+++ gcsio/src/main/java/com/google/cloud/hadoop/gcsio/GoogleCloudStorageReadOptions.java
@@ -43,7 +43,7 @@ public abstract class GoogleCloudStorageReadOptions {
   public static final boolean DEFAULT_SUPPORT_GZIP_ENCODING = true;
   public static final long DEFAULT_INPLACE_SEEK_LIMIT = 8 * 1024 * 1024;
   public static final Fadvise DEFAULT_FADVISE = Fadvise.SEQUENTIAL;
-  public static final int DEFAULT_MIN_RANGE_REQUEST_SIZE = 2 * 1024 * 1024;
+  public static final long DEFAULT_MIN_RANGE_REQUEST_SIZE = 2 * 1024 * 1024;
   public static final boolean GRPC_CHECKSUMS_ENABLED_DEFAULT = false;
   public static final long DEFAULT_GRPC_READ_TIMEOUT_MILLIS = 20 * 60 * 1000;
   public static final long DEFAULT_GRPC_READ_METADATA_TIMEOUT_MILLIS = 60 * 1000;
@@ -103,7 +103,7 @@ public abstract class GoogleCloudStorageReadOptions {
   public abstract Fadvise getFadvise();
 
   /** See {@link Builder#setMinRangeRequestSize}. */
-  public abstract int getMinRangeRequestSize();
+  public abstract long getMinRangeRequestSize();
 
   /** See {@link Builder#setGrpcChecksumsEnabled}. */
   public abstract boolean isGrpcChecksumsEnabled();
@@ -198,7 +198,7 @@ public abstract class GoogleCloudStorageReadOptions {
      * Sets the minimum size of the HTTP Range header that could be set in GCS request when opening
      * new stream to read an object.
      */
-    public abstract Builder setMinRangeRequestSize(int size);
+    public abstract Builder setMinRangeRequestSize(long size);
 
     /**
      * Sets whether to validate checksums when doing gRPC reads. If enabled, for sequential reads of
